+++ tests/end2end/features/plugins/test_plugins_bdd.py
@@ -4,7 +4,27 @@
 # Copyright 2017-2019 Christian Karl (karlch) <karlch at protonmail dot com>
 # License: GNU GPL v3, see the "LICENSE" and "AUTHORS" files for details.
 
+import imghdr
+from contextlib import suppress
+
 import pytest_bdd as bdd
+
+from vimiv import plugins
 
 
 bdd.scenarios("plugins.feature")
+
+
+@bdd.when(bdd.parsers.parse("I load the {name} plugin with {info}"))
+def load_plugin(name, info):
+    plugins._load_plugin(name, info, plugins._app_plugin_directory)
+
+
+@bdd.then(bdd.parsers.parse("The {name} format should be supported"))
+def check_format_supported(name):
+    for func in imghdr.tests:
+        with suppress(IndexError):
+            format_name = func.__name__.split("_")[-1]
+            if format_name == name:
+                return
+    assert False, f"Image format {name} is not supported"